+++ src/index.php
@@ -2,23 +2,29 @@
 require 'config.php';
 session_start();
 ob_start();
+
+// Recherche de l'utilisateur actuel dans la collection des users pour obtenir son rôle
+if (isset($_SESSION['user'])) {
+    $currentUser = $_SESSION['user'];
+    $userRole = $users->findOne(['user' => $currentUser]);
+    $currentUserRole = $userRole['role'] ?? '';
+} else {
+    $currentUserRole = '';
+}
+
 $tweets = $collection->find([], [
     'sort' => ['timestamp' => -1], // Tri par date décroissante
 ]);
 
 if (isset($_SESSION['user'])) {
-    // si session active, afficher username
     $userName = $_SESSION['user'];
     ?>
-
     <div class="my-5 text-end">
         <p>Bonjour, <span class="fw-bold fs-4"><?= $userName ?></span></p>
         <a href="log_out.php" class="btn btn-sm btn-primary mt-3">Se déconnecter</a>
     </div>
-
     <?php
 } else {
-    // Session non active, propose de se connecter
     ?>
     <div class="my-5 text-end">
         <a href="x.php" class="btn btn-sm btn-primary mt-3">Se connecter</a>
@@ -41,22 +47,25 @@ if (isset($_SESSION['user'])) {
         <p class='px-3'><?= $tweet['message'] ?></p>
         <div class='d-flex justify-content-between pb-3 px-1'>
             <form action='like_tweet.php?id=<?= $tweet['_id'] ?>' method='POST'>
-                <button class='btn btn-outline-secondary'><?php echo isset($tweet['likes']) ? $tweet['likes'] : 0; ?><i class="fa-regular fa-thumbs-up mx-1"></i></button>
+                <button class='btn btn-outline-secondary'><?= $tweet['likes'] ?? 0 ?><i class="fa-regular fa-thumbs-up mx-1"></i></button>
             </form>
             <div>
-                <p class='text-end fw-lighter fst-italic'><?= $tweet['timestamp']->toDateTime()->format('Y-m-d H:i:s') ?></p>
-                <a href='' data-bs-toggle='modal' data-bs-target='#updateModal'>Modifier</a>
+                <p class='text-end fw-lighter fst-italic'><?= $tweet['timestamp']->toDateTime()->format('H:i - m/d/Y') ?></p>
+                <?php if (isset($_SESSION['user']) && $_SESSION['user'] === $tweet['user']): ?>
+                    <a href='' data-bs-toggle='modal' data-bs-target='#updateModal<?= $tweet['_id'] ?>'>Modifier</a>
+                <?php endif ?>
             </div>
         </div>
-        <a class='deleteCross' href='delete_tweet.php?id=<?= $tweet['_id'] ?>'>❌</a>
+        <?php if ((isset($_SESSION['user']) && $_SESSION['user'] == $tweet['user']) || $currentUserRole == 'moderator'): ?>
+            <a class='deleteCross' href='delete_tweet.php?id=<?= $tweet['_id'] ?>'>❌</a>
+        <?php endif; ?>
     </div>
-<?php endforeach; ?>
 
-    <div class='modal fade' id='updateModal' tabindex='-1' aria-labelledby='updateModalLabel' aria-hidden='true'>
+    <div class='modal fade' id='updateModal<?= $tweet['_id'] ?>' tabindex='-1' aria-labelledby='updateModalLabel' aria-hidden='true'>
         <div class='modal-dialog'>
             <div class='modal-content'>
                 <div class='modal-header'>
-                    <h1 class='modal-title fs-5' id='updateModalLabel'>Modifier un tweet</h1>
+                    <h2 class='modal-title fs-5' id='updateModalLabel'>Modifier un tweet</h2>
                     <button type='button' class='btn-close' data-bs-dismiss='modal' aria-label='Close'></button>
                 </div>
                 <div class='modal-body'>
@@ -74,9 +83,8 @@ if (isset($_SESSION['user'])) {
             </div>
         </div>
     </div>
+<?php endforeach;
 
-
-<?php
 $title = 'Mini X';
 $content = ob_get_clean();
 require 'layout.php';
